Initial card rendering
This patch gets some code running and cards rendered to the screen.

Sneptile is used to generate panel data for each card.

Cards are randomly selected to be displayed at the bottom of the screen.

save.c and rng.c have been imported from Snepzhen Solitaire to help
ensure we don't get the same deal each time the game is launched.