Initial card-slide animation
This patch provides an initial card sliding animation to be used when
dealing and playing cards.

Note however, for some cards the sprite appears corrupt. This is because
the sprite table can only represent patterns 255 and below. Currently
the cards take up around 300 patterns.

To fix this, some space could be reserved in a low-address area of VRAM,
which we could populate with the single card we're about to animate.