Draw cards into a reserved area before rendering as sprites
Sprites only allow using the first 256 patterns.

As we have more card patterns than will fit in that space, we instead
reserve an area near the start of VRAM large enough for a single unique
card.

When we go to draw a card as sprites, we first copy the patterns into
this reserved area, then use only it for the sprite pattern indexes.