Initial text rendering for panels
The panel fields can now be updated with any value from 0-999.

This is done by reading in the base version of the tile and drawing over
it with a 3x5 digit font. The number is centred and can span two tiles.

Currently, the text is all white as yellow is not yet in the palette. To
fix this, a change may first be needed to Sneptile to allow pre-filling
the background palette.