Improve card image centering, add artwork for Saboteur card
As there isn't room on the card for meaningful description text, the
images can be a little lower.

Artwork for one additional card, Saboteur, is added.

This reaches the limit for how many patterns the VDP can hold at once,
without squeezing them into unused parts of other tables.

Before more card artwork is added, some kind of memory management will
be needed for the VDP to load and unload patterns as needed.