Only put cards into VRAM as they are needed
Without trampling or trying to interleave with other data in the VRAM,
we have room for only 448 patterns. Adding any more card artwork would
push past that limit and start corrupting the name table.

So, to avoid this problem, the card artwork doesn't get loaded into
VRAM at startup any more. Instead, a 24-pattern area is reserved for
each of nine dynamic cards (216 patterns), plus we already have a area
for the sliding card, and another area for the card-back.

This will let development continue, but is not yet ideal. There is no
de-duplication in these card-buffers in VRAM, so some patterns may be
duplicated many times. There is also the assumption that the game is in
the player vs. computer mode; where only one player's hand is shown.

Some future improvements that should be considered:
 - De-duplication of tiles across the full buffer to free up more VRAM
 - Pre-loading patterns for the next card to be drawn
 - Buffering more cards for modes where both hands are visible
 - Using the same buffer and already-loaded patterns  for the slide
   animation instead of re-loading into a dedicated area
 - Avoid passing both 'slot' and coordinates to render_card_as_background

For now at least, drawing the remaining card artwork can be continued.