Use card_buffer for sprites and card-backs
Previously, three separate buffers were used:
 - Card-back
 - Card-as-sprite
 - Nine-card buffer

These have now been consolidated into a ten-card buffer:
 0-7: Cards in hand
   8: Most recent discard
   9: Card back

All of the patterns (240) can be indexed below 256, so there is no need
to have a separate low-address buffer for the card-as-sprite.

This frees up 24 tiles in VRAM.

Note: The cache of which cards are in the buffer has been moved out of
function scope. This ran into a compiler bug where the cache was not
always updated.