Add fanfare sound
This sound needs to be split over two banks as it takes around 31 KiB.

We never actually use the sizes, so strip them from the generated
header files with sed. This is required at least for the first 16 KiB
block of samples as the data array will use the entire bank on its own.