Include hints on cards about their effect.
While there isn't room for a full description, at least a number can
indicate the size of the effect.

This increases the size of the card pattern data, which no-longer fits
alongside the other in-game patterns in bank_3. So, the card artwork is
now stored alone in bank_4. The sound banks have been shuffled along by
one.

The 'fast build' turned out to be a painful default, as files are not
recompiled when the palette changes or when headers change. So the old
behaviour is restored as the default (recompile all source files).

A parameter can be used to do a 'fast' build, which will only re-build
individual source files that contain changes, to speed up development
cycles. Note that editing a used header file will not trigger a re-build
in fast mode.