Return to the menu with soft-reset, improve game screen setup
The reset button on the SMS-1 will un-deal the current on-screen hand and
then return to the main menu to allow the settings to be changed and the
game restarted.

As part of this, with the game screen now being enterable more than once,
most of the setup is moved into game_start, rather than main.

Changes are made to blank the screen while drawing the game screen, so
that it pops into existence at once, rather than seeing parts of it show
up at a time over a few frames.

To account for the possibility of a game restart, the _init functions for
various elements (cards, panels, castles, etc) now reset the cache values
to force a re-draw on the first update.

RNG is seeded after the title screen, before the game is dealt. This will
allow the time spent at the title/menu to affect the R register.